+++ mac/Keyman4MacIM/Keyman4MacIM/KMPackage.m
@@ -9,20 +9,40 @@
 #import <Foundation/Foundation.h>
 #import "KMPackage.h"
 
-
 @implementation KMPackage
 
 const NSInteger kUnexpectedFileAsscociationType = 42;
+BOOL _installingKmp = NO;
+NSString *filename = nil;
 
 + (BOOL)autosavesInPlace {
     return NO;
 }
 
+// This method gets called whenever a user double-clicks a KMP file, regardless of whether
+// they decide to install it or cancel the installation. It also gets called one time up-front
+// when Keyman first loads.
 - (void)makeWindowControllers {
-    if ([self.AppDelegate debugMode])
-        NSLog(@"Attempting to display configuration window...");
-    
-    [[self.AppDelegate configWindow] showWindow:nil];
+    // The first part of this conditional ensures that the config window is created,
+    // even if we are not going to display it. If this doesn't happen, then when the
+    // user tries to open the config window manually via the menu, it won't work.
+    if ([self.AppDelegate configWindow] != nil && _installingKmp) {
+        if ([self.AppDelegate debugMode])
+            NSLog(@"Attempting to display configuration window...");
+        
+        [[self.AppDelegate configWindow] showWindow:nil];
+        _installingKmp = NO;
+    }
+    if (filename != nil) { // Should be true every time except maybe the very first time (when loading).
+        // Since we aren't really opening this file as a document, after reading it (or deciding not to)
+        // we need to close this document so the controller doesn't hold onto it. Otherwise, if the user
+        // double-clicks the same KMP file a second time it will just assume we want to open the window
+        // to display the already open document, rather than attempting to read it again.
+        if ([self.AppDelegate debugMode])
+            NSLog(@"Closing document to release presenter...");
+        [self close];
+        filename = nil;
+    }
 }
 
 - (KMInputMethodAppDelegate *)AppDelegate {
@@ -38,7 +58,7 @@ const NSInteger kUnexpectedFileAsscociationType = 42;
 
 -(BOOL)readFromFileWrapper:(NSFileWrapper *)fileWrapper ofType:(NSString *)typeName error:(NSError * _Nullable __autoreleasing *)outError {
     
-    NSString *filename = [fileWrapper filename];
+    filename = [fileWrapper filename];
     
     if ([self.AppDelegate debugMode])
         NSLog(@"readFromFileWrapper called with file: %@", filename);
@@ -54,7 +74,12 @@ const NSInteger kUnexpectedFileAsscociationType = 42;
         return NO;
     }
     
-    if (![self userConfirmsInstallationOfPackageFile:filename]) {
+    // If the following is set to NO, then it will prevent displaying the config window when
+    // makeWindowControllers gets called. (Unfortunately, something about the document infrastructure
+    // in macOS will cause it to be displayed if it has not yet been displayed, even if the user cancels
+    // the installation. This is a bug that should be fixed, but no solution is yet apparent.)
+    _installingKmp = [self userConfirmsInstallationOfPackageFile:filename];
+    if (!_installingKmp) {
         if ([self.AppDelegate debugMode])
             NSLog(@"Keyman Package file installation cancelled by user.");
         return YES; // Returning NO causes the system to report a failure to the user.
@@ -62,8 +87,9 @@ const NSInteger kUnexpectedFileAsscociationType = 42;
     
     NSString *tempFile = [self pathForTemporaryKmpFile:filename];
     NSURL *fileURL = [[NSURL alloc] initFileURLWithPath:tempFile];
-    if (![fileWrapper writeToURL:fileURL options:NSFileWrapperWritingAtomic originalContentsURL:nil error:outError])
+    if (![fileWrapper writeToURL:fileURL options:NSFileWrapperWritingAtomic originalContentsURL:nil error:outError]) {
         return NO;
+    }
     
     BOOL didUnzip = [self.AppDelegate unzipFile:tempFile];
     
@@ -88,6 +114,12 @@ const NSInteger kUnexpectedFileAsscociationType = 42;
     if ([self.AppDelegate debugMode])
         NSLog(@"Asking user to confirm installation...");
     
+    // Attempted the following (with and without the first line) to try to
+    // get Keyman to be active foreground app so the alert message would not
+    // appear behind other window(s). Unfortunately, it doesn't work, probably
+    // because of Keyman being an input method.
+//    [[self.AppDelegate configWindow] showWindow:nil];
+//    [[NSRunningApplication currentApplication] activateWithOptions:0];
     BOOL result = [alert runModal] == NSAlertFirstButtonReturn;
     return result;
 }